arm: Don't reject early mov?fcc patterns that we might be able to handle
The define_expand patterns for movdfcc, movsfcc and movhfcc had overly
tight contstraints that could cause the compiler to reject these
patterns when re-ordering the operands could lead to a successful
match.  Relax the initial predicate test and rely on the test after
arm_validize_comparison has run to determine whether this is something
we can support.  This fixes some test failures which were introduced
in the fix for PR118460

gcc/ChangeLog:

	PR target/118460
	* config/arm/arm.md (movhfcc): Use expandable_comparison_operator.
	(movsfcc, movdfcc): Likewise.