sccp: Fix order of removal of phi (again) [PR122599]
This time we are gimplifying the expression and call
fold_stmt during the gimplification (which is fine) but
since we removed the phi and the expression references ssa
names in the phi indirectly, things just fall over inside the ranger.

This moves the removal of the phi until gimplification happens as it
might refer back to the ssa name that the phi defines.

Pushed as obvious after bootstrap test on x86_64-linux-gnu.

	PR tree-optimization/122599

gcc/ChangeLog:

	* tree-scalar-evolution.cc (final_value_replacement_loop): Move
	the removal of the phi until after the gimplification of the final
	value expression.

gcc/testsuite/ChangeLog:

	* gcc.dg/torture/pr122599-1.c: New test.

Signed-off-by: Andrew Pinski <andrew.pinski@oss.qualcomm.com>