mergephi: Remove worklist
Since the worklist was never added to and the anlysis part can benifit
from the work part, we can combine the analayis part with the work part.
This should get a small speedup for this pass

Looking into the history here, remove_forwarder_block used to add to the worklist
but remove_forwarder_block_with_phi never did.

This is the first step in moving part of the functionality of mergephi into
cfgcleanup.

gcc/ChangeLog:

	* tree-cfgcleanup.cc (pass_merge_phi::execute): Remove worklist.

Signed-off-by: Andrew Pinski <andrew.pinski@oss.qualcomm.com>