ada: Avoid non-serious errors masking fatal errors
Is_Redundant_Error_Message is used to filter diagnostic messages
that would appear on the same line to avoid the noise comming from
cascading error messages. However we still want to trigger an error
even if the line already had a warning or a non-serious error at the
same location so that we now that a fatal error has occured and the
compiler knows how to exit correctly in that scenario.

gcc/ada/ChangeLog:

	* erroutc.adb (Is_Redundant_Error_Message): Avoid non-serious errors
	masking fatal errors.