amdgcn: Consolidate mkoffload setup constructors
We don't need every mkoffload runtime setting to use it's own constructor.
There was only two committed, but I have more uses for this soon.  In theory,
we could also use this setup to choose not to register the kernel with libgomp.

The behaviour is not changed, just the generated code structure.

gcc/ChangeLog:

	* config/gcn/mkoffload.cc (process_asm): Replace "configure_stack_size"
	constructor with a new regular function, "mkoffload_setup".
	(process_obj): Call mkoffload_setup from the "init" constructor.