Merge remove_forwarder_block_with_phi into remove_forwarder_block
This is the last cleanup in this area. Merges the splitting functionality
of remove_forwarder_block_with_phi into remove_forwarder_block.
Now mergephi still has the ability to split the edges when merging the forwarder
block with a phi. But this reduces the non-shared code a lot.

gcc/ChangeLog:

	* tree-cfgcleanup.cc (tree_forwarder_block_p): Remove must argument.
	(remove_forwarder_block): Add can_split
	argument. Handle the splitting case (iff phis in bb).
	(cleanup_tree_cfg_bb): Update argument to tree_forwarder_block_p.
	(remove_forwarder_block_with_phi): Remove.
	(pass_merge_phi::execute): Update argument to tree_forwarder_block_p
	and call remove_forwarder_block instead of remove_forwarder_block_with_phi.

Signed-off-by: Andrew Pinski <andrew.pinski@oss.qualcomm.com>