gccrs: Fmt: Simplify pragma diagnostic setup
There was a typo in the original commit where the diagnostic context was not being popped,
however popping the context still causes issues while bootstrapping.

For now, just mark -Warray-bounds as a warning to fix bootstrap on trunk and think about
a proper fix (probably adding a push and pop on every file including rust-fmt.h...) later if
the C++ warning issue is still not fixed.

gcc/rust/ChangeLog:

	* ast/rust-fmt.h: Simplify diagnostic avoidance.