ada: Fix actual parameters in call
A recent patch made Multi_Module_Symbolic_Traceback have two consecutive
formal parameters of type Boolean, which opens the door for mixing up
actual parameters in calls. And that mistake was actually made in a call
introduced by the same patch.

This commit fixes the call and also introduces a new enumerated type to
make this kind of mistake less likely in the future.

gcc/ada/ChangeLog:

	* libgnat/s-dwalin.ads (Display_Mode_Type): New enumerated type.
	(Symbolic_Traceback): Use new type in profile.
	* libgnat/s-dwalin.adb (Symbolic_Traceback): Use new type in profile
	and adapt body.
	* libgnat/s-trasym__dwarf.adb (Multi_Module_Symbolic_Traceback): Fix
	wrong call in body of one overload. Use new type in profile. Adapt
	body.
	(Symbolic_Traceback, Symbolic_Traceback_No_Lock,
	Module_Symbolic_Traceback): Use new type in profile and adapt body.
	(Calling_Entity): Adapt body.