cfgloop: Modify loop_exits_{to,from}_bb_p return type to edge
Given that when finding whether the predicate in question is satisfied
or not we already do the heavy-lifting of identifying the specific
edge that matches the particular criterion, it is wasteful to throw
the edge information away, only to potentially have to recalculate it
when true is returned.

Rather, given the ability of treating a valid pointer as true and,
conversely, the NULL pointer as false, we can return the edge for
should we wish to use it, while keeping the function's existing calls
in the code as is.

gcc/ChangeLog:

	* cfgloop.cc (loop_exits_to_bb_p): Change return type.
	(loop_exits_from_bb_p): Likewise.
	* cfgloop.h: (loop_exits_to_bb_p): Likewise.
	(loop_exits_from_bb_p): Likewise.