[vxworks] wrap base/b_NULL.h to override NULL
Some versions of vxworks define NULL to __nullptr in C++, assuming
C++11, which breaks at least a number of analyzer tests that get
exercised in C++98 mode.

Wrap the header that defines NULL so that, after including it, we
override the NULL definition with the one provided by stddef.h.

That required some infrastructure to enable subdirectories in extra
headers.  Since USER_H filenames appear as dependencies, that limits
the possibilities or markup, so I went for a filesystem-transparent
sequence that doesn't appear in any extra_headers whatsoever, namely
/././, to mark the beginning of the desired install name.

Co-Authored-By: Olivier Hainque <hainque@adacore.com>

for  gcc/ChangeLog

	* config/vxworks/base/b_NULL.h: New.
	* config.gcc (extra_headers) <*-*-vxworks*>: Add it.
	* Makefile.in (stmp-int-hdrs): Support /././ markers in USER_H
	to mark the beginning of the install name.  Document.
	* doc/sourcebuild.texi (Headers): Document /././ marker.