middle-end: fix missing converts in vec_cond_expr simplification [PR122724]
The following avoids type inconsistencies in .COND_op generated by
simplifications of VEC_COND_EXPRs.

This is a backport of r14-5317-g20aa06490ab57da7729a24bae7c4ec2f5918ec91
but with a testcase that triggered an ICE on the 13 branch (with the
same root cause as the original PR).  It isn't an exact cherry-pick
because some of the patterns that were patched in the original fix don't
exist in GCC 13.

gcc/ChangeLog:

	PR tree-optimization/112469
	PR tree-optimization/122724
	* match.pd (cond ? op a : b -> .COND_op (cond, a, b)): Add
	missing view_converts.

gcc/testsuite/ChangeLog:

	PR tree-optimization/112469
	PR tree-optimization/122724
	* gcc.target/aarch64/sve/pr122724.c: New test.

Co-Authored-By: Alex Coplan <alex.coplan@arm.com>