vect: Make SELECT_VL a convert optab.
Currently select_vl is a direct optab with its mode always Xmode/Pmode.
This does not give us sufficient freedom to enable/disable vsetvl
(=SELECT_VL) depending on the vector mode.

This patch makes select_vl a convert optab and adjusts the associated IFN
functions as well as the query/emit code in the vectorizer.

With this patch nothing new is actually exercised yet.  This is going to
happen in a separate riscv patch that enables "VLS" select_vl.

gcc/ChangeLog:

	* config/riscv/autovec.md (select_vl<mode>): Rename to...
	(select_vl<V:mode><P:mode>): ...this.
	* doc/md.texi: Document new behavior.
	* internal-fn.cc (select_vl_direct): Make
	(expand_select_vl_optab_fn): Adjust for convert optab.
	(direct_select_vl_optab_supported_p): Ditto.
	* internal-fn.def (SELECT_VL): Ditto.
	* optabs.def (OPTAB_CD): Add select_vl.
	(OPTAB_D): Remove select_vl.
	* tree-vect-loop-manip.cc (vect_set_loop_controls_directly):
	Adjust for convert select_vl optab.
	* tree-vect-loop.cc: Ditto.